requirements.txt
Django
djangorestframework
numpy 
openpyxl
pandas 